Use Tcl_QueueEvent instead of Tcl_CreateTimeHandler to reschedule channel event handlers.
The channel subsystem should be able to drive I/O without resorting to the
timer event source.  One relevant Tcl report is [ef28eb1f15167156c898].  This
is a step in the direction of accomplishing that.  Investigation of the changes
needed to correctly manage reference counting of Channel and ChannelState data
structures revealed several deficiencies in the current approach.  For one
thing, Tcl_Preserve/Tcl_Release/Tcl_EventuallyFree were used, but these are not
efficient, and not intended to be used as extensively as they were intclIO.c.
First, they loop over a linked list of pointers under their management.
Second, they employ a (slow) mutex, which is not necessary since interpreters
that share a channel are confined to a single thread.  A more simple
alternative is now in place:  Since channels can be stacked, the channel state
can be freed when the bottom channel is freed.  The bottom channel in turn can
be freed when it is closed and when the reference count drops to zero.

The new approach of using Tcl_QueueEvent() to enqueue a new event channel
handler lead to Tcl_DoOneEvent() to never poll event sources since the queue
never became empty.  Therefore, Tcl_DoOneEvent() now marks the last event on
the queue after all event sources get a chance to enqueue events. After it gets
to and processes that marked event, Tcl_DoOneEvent() processes one event from
the idle queue.  This makes the idle queue more woven into normal event
processing, but at a rate that should be reasonable.  This change should also
alleviate existing situations where the idle queue is overstarved, and perhaps
even obsolete [update idletasks].  One caveat is that it is still possible to
starve the idle event queue by continuously placing events at the front of the
queue.  Doing that is discouraged and might even be made impossible in the
future.